chat_config_base
